requirements.txt
absl-py==0.10.0
addict==2.4.0
astor==0.8.1
astunparse==1.6.3
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
block-extractor-cuda==0.0.0
cachetools==4.1.1
certifi==2019.11.28
cffi==1.11.5
chardet==3.0.4
click==8.0.3
cloudpickle==1.2.1
count-get-cuda==0.0.0
cpools==0.0.0
cryptography==2.8
cycler==0.10.0
Cython==0.28.5
cytoolz==0.10.0
dask==2.1.0
decorator==4.4.0
defusedxml==0.5.0
dill==0.3.3
dismap-get-cuda==0.0.0
dnspython==1.16.0
dominate==2.4.0
ecdsa==0.13
entrypoints==0.3
eventlet==0.25.1
fastrlock==0.8
Flask==2.0.2
future==0.17.1
gast==0.2.2
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==0.4.15
grpcio==1.23.0
h5py==2.10.0
idna==2.8
imageio==2.5.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==2.0.0
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==2.0.1
jedi==0.15.1
Jinja2==3.0.3
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.31.0
local-attn-reshape-cuda==0.0.0
lpips==0.1.3
Markdown==3.3
MarkupSafe==2.0.1
matplotlib==3.0.2
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
mmcv==1.3.9
monotonic==1.5
natsort==7.0.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
neuralgym==0.0.1
notebook==6.0.1
numba==0.48.0
numpy==1.19.2
nvidia-dali==0.13.0
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.1.0.25
opencv-python==4.4.0.46
opt-einsum==3.3.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycparser==2.18
pycryptodome==3.9.4
Pygments==2.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.1
pyrsistent==0.15.4
PySocks==1.7.1
python-dateutil==2.8.0
pytorch-fid==0.2.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.3.1
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.24.0
requests-oauthlib==1.3.0
resample2d-cuda==0.0.0
rsa==4.6
scikit-image==0.14.2
scikit-learn==0.19.1
scipy==1.2.0
Send2Trash==1.5.0
Shapely==1.7.1
six==1.12.0
TBB==0.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorboardX==1.8
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
torch==1.2.0
torchfile==0.1.0
torchvision==0.2.2
tornado==6.0.3
tqdm==4.19.9
traitlets==4.3.2
urllib3==1.25.10
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yacs==0.1.8
yapf==0.31.0
zipp==3.3.0
